First steps to enable the picture in the toolbar
